Rename to s-select. Added tests, now 100% covered Removed IE8 support

+++ addon/components/s-select.js
@@ -1,21 +1,23 @@
 import Ember from 'ember';
-import layout from '../templates/components/x-select';
+import layout from '../templates/components/s-select';
 
 const {
   Component,
   computed,
   get,
-  inject,
   isBlank,
   isPresent,
-  run
+  run,
+  isArray
 } = Ember;
 
-const isEdgeIe = typeof StyleMedia !== 'undefined';
+const containsGroup = list => {
+  return isArray(list) && list.find(ele => ele.parentId);
+};
 
 export default Component.extend({
   layout,
-  classNames: ['ember-select'],
+  classNames: ['s-select'],
   classNameBindings: [
     'isOpen:es-open', 'isFocus:es-focus',
     'canSearch::es-select', 'multiple:es-multiple'
@@ -25,7 +27,6 @@ export default Component.extend({
   canSearch: true,
   disabled: false,
   dropdown: 'select-dropdown',
-  freeText: false,
   isDirty: false,
   isFocus: false,
   isOpen: false,
@@ -38,8 +39,6 @@ export default Component.extend({
   labelKey: 'label',
   valueKey: 'value',
 
-  messageBus: inject.service(),
-
   canClear: computed.and('enabled', 'canSearch', 'hasOptions'),
   canOpen: computed.or('hasInput', 'openOnFocus'),
   enabled: computed.not('disabled'),
@@ -50,6 +49,7 @@ export default Component.extend({
   hasValues: computed.notEmpty('values'),
   multiple: computed.bool('values'),
   shouldFilter: computed.or('isDirty', 'multiple', 'hasChanged'),
+  hideDropdown: computed.not('isOpen'),
 
   input: computed(function() {
     return document.querySelector(`#${this.elementId} input`);
@@ -67,7 +67,6 @@ export default Component.extend({
 
   init() {
     this._super(...arguments);
-
     if (this.disabled) {
       this.set('canSearch', false);
     }
@@ -77,17 +76,9 @@ export default Component.extend({
     }
 
     this.set('oldValue', this.get('value'));
-  },
 
-  didReceiveAttrs() {
-    this._super(...arguments);
-
-    /* IE10+ Triggers an input event when focus changes on
-     * an input element if the element has a placeholder.
-     * https://connect.microsoft.com/IE/feedback/details/810538/
-     */
-    if (document.documentMode) {
-      this.set('placeholder', null);
+    if (containsGroup(this.get('model'))) {
+      this.set('dropdown', 'select-dropdown-group');
     }
   },
 
@@ -113,9 +104,7 @@ export default Component.extend({
     let oldValue = this.get('oldValue');
     if (oldValue && newValue && oldValue !== newValue) {
       let { label } = this.retrieveOption(newValue);
-      if (label !== this.get('token')) {
-        this.setOption(newValue);
-      }
+      label !== this.get('token') && this.setOption(newValue);
     }
 
     this.set('oldValue', newValue);
@@ -125,8 +114,7 @@ export default Component.extend({
     blur() {
       if (this.get('isDirty')) {
         // Clear unallowed input in strict single mode
-        let option = this.get('freeText') ? this.get('token') : '';
-        this.setOption(option, false, !this.get('multiple'));
+        this.setOption('', false, !this.get('multiple'));
       }
 
       this.setProperties({
@@ -159,16 +147,10 @@ export default Component.extend({
     },
 
     focus() {
-      if (this.get('disabled')) {
-        return this.send('blur');
-      }
-
       let input = this.get('input');
-      if (input) {
-        input.focus();
-      }
 
-      if (input && !this.get('isFocus') && this.get('canSearch')) {
+      input.focus();
+      if (!this.get('isFocus') && this.get('canSearch')) {
         // Select text (similar to TAB)
         input.select();
       }
@@ -180,28 +162,18 @@ export default Component.extend({
 
     keypress(e) {
       let isOpen = this.get('isOpen');
+      this.set('keyEvent', e);
 
-      switch (e.keyCode) {
+      switch (e.which) {
         case 8: { // Backspace
           let values = this.get('values');
           if (isPresent(values) && this.get('token') === '') {
-            let last = this.getElement(values, get(values, 'length') - 1);
-            this.attrs.onRemove(last);
+            let removedVal = this.get('values').pop();
+            this.attrs.onRemove && this.attrs.onRemove(removedVal);
             e.preventDefault();
           }
-
           break;
         }
-
-        case 9: // TAB
-        case 13: // Enter
-          if (isOpen) {
-            this.get('messageBus').publish('select-key', e);
-          } else if (this.get('freeText')) {
-            this.send('select', this.get('token'), false);
-          }
-
-          break;
         case 27: // ESC
           if (this.get('canSearch') && this.get('hasInput')) {
             this.send('clear');
@@ -211,43 +183,25 @@ export default Component.extend({
           break;
         case 38: // Up Arrow
         case 40: // Down Arrow
-          if (isOpen) {
-            this.get('messageBus').publish('select-key', e);
-          } else {
+          if (!isOpen) {
             this.set('isOpen', true);
           }
-
           e.preventDefault();
           break;
       }
     },
 
     remove(selection) {
-      this.attrs.onRemove(selection);
+      let values = this.get('values');
+      let index = values.indexOf(selection);
+      values.splice(index, 1);
+      this.attrs.onRemove && this.attrs.onRemove(selection);
       this.send('focus');
     },
 
     select(option, selected) {
-      let isNew = !selected && this.get('freeText') && this.get('isDirty');
-      let allowNew = isPresent(this.attrs.onCreate);
       let valid = isPresent(option);
-
-      /* Notify when option is either
-       *  - selected
-       *  - new, empty and cannot be created */
-      let notify = selected || isNew && !allowNew;
-
-      if (allowNew && valid && isNew) {
-        this.attrs.onCreate(option);
-      }
-
-      this.setOption(option, selected, notify);
-
-      /* Blur on selection when single
-       * IE & Edge do not run the events in proper order */
-      if (!this.get('multiple') && !isEdgeIe) {
-        this.get('input').blur();
-      }
+      valid && this.setOption(option, selected, true);
     }
   },
 
@@ -302,7 +256,7 @@ export default Component.extend({
       this.set('token', label);
     }
 
-    // Ensure the component hasn't been destroyed before updating
+    // Ensure the component is DOM ready before updating
     let input = this.get('input');
     if (input) {
       input.value = label;
